platform/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-bigquery==3.25.0
pydantic==2.9.0
python-multipart==0.0.18
PyYAML==6.0.1
protobuf==5.29.5
requests==2.32.4
urllib3==2.6.3
google-generativeai
